debug = true
Have you ever played the hit game Among Us? In this viral social deduction game first released by the American game development company Innersloth in 2020, 5-15 players are given roles, either "Crewmates", or "Impostors". The goal of each crewmate is to complete designated tasks around one of the game's three game maps while avoiding the Impostors. The goal of an Impostor is to kill each crewmate without being discovered. Players can report a dead body they might find, or call and Emergency Meeting to gather all other players and call a vote. Each player deliberates and communicates during these meetings and chooses who they believe is most likely the Impostor. The player with the most votes is "ejected" from the spaceship, the setting of each Among Us map. This player is removed from the game. If all of the Impostors are voted out of the game, the Crewmates win, but Impostors can also steal the victory by killing enough Crewmates to never be voted out.